LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/fill-and-stroke-styles/2d.pattern.image.zeroheight-expected.txt

2d.pattern.image.zeroheight
Actual output:

FAIL Canvas test: 2d.pattern.image.zeroheight assert_equals: ctx.createPattern(img, 'repeat') === null (got [object CanvasPattern][object], expected [object]) expected null but got object "[object CanvasPattern]"
